@@ -5,29 +5,29 @@
   "license": "MIT",
   "private": true,
   "dependencies": {
-    "@fortawesome/fontawesome-svg-core": "^1.2.19",
-    "@fortawesome/free-brands-svg-icons": "^5.9.0",
-    "@fortawesome/free-solid-svg-icons": "^5.9.0",
-    "@fortawesome/react-fontawesome": "^0.1.4",
-    "@types/node": "^14.14.6",
-    "@types/react": "^17.0.3",
-    "@types/react-dom": "^17.0.2",
-    "@types/react-responsive": "^8.0.2",
-    "@types/react-router-dom": "^5.1.0",
-    "@types/react-transition-group": "^4.2.3",
-    "@types/socket.io-client": "^1.4.34",
-    "@types/uuid": "^8.3.0",
-    "axios": "^0.21.0",
-    "react": "^17.0.1",
-    "react-dom": "^17.0.1",
-    "react-responsive": "^8.1.1",
-    "react-router-dom": "^5.0.1",
-    "react-scripts": "^4.0.0",
-    "react-transition-group": "^4.2.1",
-    "socket.io-client": "^4.0.0",
-    "ts-node": "^9.0.0",
-    "typescript": "^4.0.5",
-    "uuid": "^8.3.1"
+    "@fortawesome/fontawesome-svg-core": "^1.2.36",
+    "@fortawesome/free-brands-svg-icons": "^5.15.4",
+    "@fortawesome/free-solid-svg-icons": "^5.15.4",
+    "@fortawesome/react-fontawesome": "^0.1.16",
+    "@types/node": "^16.11.7",
+    "@types/react": "^17.0.34",
+    "@types/react-dom": "^17.0.11",
+    "@types/react-responsive": "^8.0.4",
+    "@types/react-router-dom": "^5.3.2",
+    "@types/react-transition-group": "^4.4.4",
+    "@types/socket.io-client": "^3.0.0",
+    "@types/uuid": "^8.3.1",
+    "axios": "^0.24.0",
+    "react": "^17.0.2",
+    "react-dom": "^17.0.2",
+    "react-responsive": "^9.0.0-beta.4",
+    "react-router-dom": "^6.0.2",
+    "react-scripts": "^4.0.3",
+    "react-transition-group": "^4.4.2",
+    "socket.io-client": "^4.3.2",
+    "ts-node": "^10.4.0",
+    "typescript": "^4.4.4",
+    "uuid": "^8.3.2"
   },
   "devDependencies": {},
   "scripts": {
