@@ -7,28 +7,27 @@
   "private": false,
   "dependencies": {
     "bcryptjs": "^2.4.3",
-    "config": "^3.2.0",
-    "dotenv": "^8.1.0",
+    "config": "^3.3.6",
+    "dotenv": "^10.0.0",
     "express": "^4.17.1",
-    "express-validator": "^6.1.1",
+    "express-validator": "^6.13.0",
     "jsonwebtoken": "^8.5.1",
-    "mongoose": "^5.6.4",
-    "nodemailer": "^6.5.0",
+    "mongoose": "^6.0.13",
+    "nodemailer": "^6.7.1",
     "path": "^0.12.7",
-    "socket.io": "^4.0.0",
-    "ts-node": "^9.0.0",
-    "typescript": "^4.0.5"
+    "socket.io": "^4.3.2",
+    "ts-node": "^10.4.0",
+    "typescript": "^4.4.4"
   },
   "devDependencies": {
     "@types/bcryptjs": "^2.4.2",
-    "@types/config": "^0.0.38",
-    "@types/express": "^4.17.9",
+    "@types/config": "^0.0.40",
+    "@types/express": "^4.17.13",
     "@types/express-validator": "^3.0.0",
-    "@types/jsonwebtoken": "^8.5.0",
-    "@types/nodemailer": "^6.4.1",
-    "@types/socket.io": "^2.1.11",
-    "concurrently": "^6.0.0",
-    "nodemon": "^2.0.6"
+    "@types/jsonwebtoken": "^8.5.6",
+    "@types/nodemailer": "^6.4.4",
+    "concurrently": "^6.4.0",
+    "nodemon": "^2.0.15"
   },
   "scripts": {
     "start": "NODE_ENV=production node build/index.js",
@@ -49,7 +48,7 @@
     "type": "git"
   },
   "engines": {
-    "node": "^15.5",
+    "node": "^16.13",
     "yarn": "^1.22"
   }
 }
