@@ -1,5 +1,5 @@
 import { FC, useReducer } from "react";
-import { Socket } from "socket.io";
+import { Socket } from "socket.io-client";
 import socketIOClient from "socket.io-client";
 
 import AppContext from "./AppContext";
